Moves navigation from connections to stripe-ui-core.
