+++ stripe-ui-core/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     implementation libs.compose.foundation
     implementation libs.compose.material
     implementation libs.compose.materialIcons
+    implementation libs.compose.navigation
     implementation libs.compose.ui
     implementation libs.compose.uiToolingPreview
     implementation libs.diskLruCache
